requirements.txt
# Core
groq>=0.9.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
scikit-learn>=1.5.0

# New — efficiency & UX
rich>=13.0.0

# Testing
pytest>=8.0.0
